refactor(live-preview): split widgets.ts monolith into 10 files under widgets/
Context
-------
extensions/live-preview/ had both a 594-line widgets.ts file with 12
widget classes AND a widgets/ folder with 9 files using the kebab-case
`<name>-widget.ts` convention. New contributors had to learn "some
widgets are in the big file, some are in the folder, why?" — the
answer being nothing more than historical accident.

Consolidate to the folder convention. While doing the split, drop
`HorizontalRuleWidget` — it's been dead since Fase 8, when the HR
case was rewritten as a `Decoration.mark` + `Decoration.line` pair
(see handlers/simple-widget-handlers.ts, horizontalRuleHandler). No
code or test references it.

Solution
--------
Ten new files, one class each, naming scheme matches the existing
widgets/ entries:
  - widgets/task-checkbox-widget.ts (TaskCheckboxWidget)
  - widgets/ordered-list-marker-widget.ts (OrderedListMarkerWidget)
  - widgets/unordered-list-marker-widget.ts (UnorderedListMarkerWidget)
  - widgets/image-widget.ts (ImageWidget)
  - widgets/audio-widget.ts (AudioWidget)
  - widgets/video-widget.ts (VideoWidget)
  - widgets/wikilink-image-embed-widget.ts (WikilinkImageEmbedWidget)
  - widgets/wikilink-note-embed-widget.ts (WikilinkNoteEmbedWidget +
    private resolveEmbedTarget + loadEmbedContent)
  - widgets/meta-bind-select-widget.ts (MetaBindSelectWidget +
    exported createMetaBindSelect — shared with TableWidget)
  - widgets/table-widget.ts (TableWidget + private renderInlineSegments
    + renderCellContent helpers)

widgets.ts deleted. The `openWikilinkTarget` re-export it carried had
zero external consumers — dropped.

Imports in 8 consumers rewritten to the per-file paths instead of the
barrel:
  - plugins/table-field.ts           → widgets/table-widget
  - plugins/meta-bind-input-plugin.ts → widgets/meta-bind-select-widget
  - plugins/audio-plugin.ts          → widgets/audio-widget
  - plugins/image-plugin.ts          → widgets/image-widget
  - plugins/video-plugin.ts          → widgets/video-widget
  - plugins/frontmatter-field.ts     → widgets/frontmatter-widget
  - plugins/wikilink-embed-plugin.ts → widgets/wikilink-{image,note}-embed-widget
  - inline/handlers/simple-widget-handlers.ts → widgets/{task-checkbox,ordered-list-marker}-widget

Each consumer now imports only what it uses — tree-shaking can prune
unreachable widgets that the barrel kept alive.

Testing
-------
Full suite: 5096 tests green. pnpm check clean.
