+++ src/lib/core/markdown-editor/extensions/live-preview/plugins/table-field.ts
@@ -2,7 +2,7 @@ import { RangeSetBuilder } from '@codemirror/state';
 import type { EditorState } from '@codemirror/state';
 import { Decoration, type DecorationSet, EditorView, ViewPlugin, type ViewUpdate } from '@codemirror/view';
 import { findAllTables } from '../parsers/table';
-import { TableWidget } from '../widgets';
+import { TableWidget } from '../widgets/table-widget';
 import { hiddenLineDeco } from '../styles';
 import { checkUpdateAction } from '../core/check-update-action';
 import { shouldShowSource } from '../core/should-show-source';
